requirements.txt
torch
torchvision
numpy
opencv-python
albumentations
matplotlib
pillow
flask
flask-cors
requests
python-dotenv
